+++ scripts/renderers.py
@@ -76,7 +76,8 @@ class CsvRenderer():
                     data_to_render.append(toggle_dict)
 
         # sort data by either annotation_name or state_name
-        sorting_key = lambda datum: datum.get("name", "")
+        # Without the `or ""` the sorting key can be None if name is set to None explicitly.
+        sorting_key = lambda datum: datum.get("name", "") or ""
         return sorted(data_to_render, key=sorting_key)
 
     def get_sorted_headers_from_toggles(self, flattened_toggles_data, initial_header=None):