+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Change Log
 
 Unreleased
 ~~~~~~~~~~
+* Handle the case where certain toggle names come in as ``None`` when generating summary reports.
 
 [4.0.0] - 2021-01-24
 ~~~~~~~~~~~~~~~~~~~~